fix: PostService의 deleteComment()와 updateComment()에서 Comment 객체를 불러오는 것이 유리하다 판단하여 불러오는 방법으로 수정

+++ src/main/java/com/hong/ForPaw/repository/PostRepository.java
@@ -37,7 +37,7 @@ public interface PostRepository extends JpaRepository<Post, Long> {
 
     @Modifying
     @Query("UPDATE Post p SET p.commentNum = p.commentNum - :decrementNum WHERE p.id = :postId AND p.commentNum > 0")
-    void decrementCommentNumById(@Param("postId") Long postId, @Param("decrementNum") Long decrementNum);
+    void decrementCommentNumById(@Param("postId") Long postId, @Param("decrementNum") Integer decrementNum);
 
     @Query("SELECT p.user.id FROM Post p WHERE p.id = :postId")
     Optional<Long> findUserIdByPostId(@Param("postId") Long postId);
